+++ CHANGELOG.md
@@ -12,7 +12,8 @@
   - `.onDestroy()` to `.onClose()`
   - `postDestroyRxCollection` to `postCloseRxCollection`
   - `preDestroyRxDatabase` to `preCloseRxDatabase`
-  - 
+- CHANGE (internal) migration-storage plugin: Remove catch from cleanup
+- CHANGE (internal) rename RX_PIPELINE_CHECKPOINT_CONTEXT to `rx-pipeline-checkpoint`
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->